+++ src/main/kotlin/com/macoder/styling/authentication/service/MemberService.kt
@@ -21,17 +21,17 @@ class MemberService(
     private val encoder: PasswordEncoder
 ) {
     @Transactional(readOnly = true)
-    fun getMemberInfo(id: Int) = MemberInfoResponse.from(memberRepository.findByIdOrThrow(id, "존재하지 않는 회원입니다."))
+    fun getMemberInfo(id: Long) = MemberInfoResponse.from(memberRepository.findByIdOrThrow(id, "존재하지 않는 회원입니다."))
 
     @Transactional
-    fun deleteMember(id: Int): MemberDeleteResponse {
+    fun deleteMember(id: Long): MemberDeleteResponse {
         if (!memberRepository.existsById(id)) return MemberDeleteResponse(false)
         memberRepository.deleteById(id)
         return MemberDeleteResponse(true)
     }
 
     @Transactional
-    fun updateMember(id: Int, request: MemberUpdateRequest): MemberUpdateResponse {
+    fun updateMember(id: Long, request: MemberUpdateRequest): MemberUpdateResponse {
         val member = memberRepository.findByIdOrNull(id)?.takeIf { encoder.matches(request.password, it.password) }
             ?: throw IllegalArgumentException("아이디 또는 비밀번호가 일치하지 않습니다.")
         member.update(request, encoder)
@@ -39,7 +39,7 @@ class MemberService(
     }
 
     @Transactional
-    fun updateMemberToStylist(id: Int, request: MemberUpdateRequest): MemberUpdateStylistResponse {
+    fun updateMemberToStylist(id: Long, request: MemberUpdateRequest): MemberUpdateStylistResponse {
         val member = memberRepository.findByIdOrNull(id)?.takeIf { encoder.matches(request.password, it.password) }
             ?: throw IllegalArgumentException("아이디 또는 비밀번호가 일치하지 않습니다.")
